refactor: update sleep_seconds to support fractional delay

+++ src/utils/cross_platform.h
@@ -1,6 +1,6 @@
 #ifndef CROSS_PLATFORM_H
 #define CROSS_PLATFORM_H
 
-void sleep_seconds(unsigned int seconds);
+void sleep_seconds(float seconds);
 
 #endif
